First Project = "Church Stats"
Church Stats is a simple web app that allows the user to (1) record
stats from a particular Sunday worship service, (2) submit their input
and (3) display their input in a separate div.

The primary goal of this project is to become more confident in HTML5,
CSS3, Javascript & Github. No framework (i.e. Bootstrap, jQuery) will be
used in this project...as though you couldn't already see that.

As of July 28, I have not learned Javascript event handling. For now a
pre-determined text will be published in the "results" container.